+ This is another major overhaul of the RECORD/REPLAY Code. The biggest change   is having single log entry per event. Here is a list of major changes:   - One log entry per synchronization event.   - As a side-effect, the XXX_return events don't exist anymore.   - Log patching not required now.   - Log annotation is done on-the-fly at the time of record itself.   - Fake replay of pthread mutexes, condition variables, and rwlocks. + This changeset is an ongoing work, but its getting very close to become   stable.

+++ dmtcp/src/ptracewrapper.cpp
@@ -53,7 +53,8 @@ extern "C" int ptrace_info_list_size() {
 }
 
 extern "C" struct ptrace_info get_next_ptrace_info(int index) {
-  if (index >= ptrace_info_list.size()) return EMPTY_PTRACE_INFO;
+  if ((unsigned int)index >= ptrace_info_list.size())
+    return EMPTY_PTRACE_INFO;
 
   dmtcp::list<struct ptrace_info>::iterator it;
   int local_index = 0;
